fix(ci): keep report identities portable and unique
